+++ src/org/iguana/generator/ide/GenerateParserFiles.java
@@ -2,14 +2,10 @@ package org.iguana.generator.ide;
 
 import org.iguana.generator.Generator;
 import org.iguana.grammar.runtime.RuntimeGrammar;
-import org.iguana.grammar.runtime.RuntimeRule;
-import org.iguana.grammar.symbol.Nonterminal;
 
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
 
 import static org.iguana.generator.GeneratorUtils.writeToJavaFile;
 import static org.iguana.utils.string.StringUtil.toFirstUpperCase;
@@ -17,10 +13,13 @@ import static org.iguana.utils.string.StringUtil.toFirstUpperCase;
 public class GenerateParserFiles extends Generator {
 
     private final String parserGenDirectory;
+    private final String className;
+    private final List<String> metaSymbolNodes = Arrays.asList("Opt", "Star", "Plus", "Group", "Alt", "Start");
 
     public GenerateParserFiles(RuntimeGrammar grammar, String grammarName, String packageName, String genDirectory) {
         super(grammar, grammarName, packageName, genDirectory);
         this.parserGenDirectory = new File(genDirectory, "parser").getAbsolutePath();
+        this.className = toFirstUpperCase(grammarName);
     }
 
     public void generate() {
@@ -28,22 +27,15 @@ public class GenerateParserFiles extends Generator {
     }
 
     private void generateIggyElementTypes() {
-        List<String> metaSymbolNodes = Arrays.asList("Opt", "Star", "Plus", "Group", "Alt", "Start");
-
-        String className = toFirstUpperCase(grammarName);
         StringBuilder sb = new StringBuilder();
         sb.append("// This file has been generated, do not directly edit this file!\n");
         sb.append("package " + grammarName.toLowerCase() + ".ide.parser;\n");
         sb.append("\n");
-        sb.append("import com.intellij.lang.ASTNode;\n");
-        sb.append("import com.intellij.psi.PsiElement;\n");
         sb.append("import com.intellij.psi.tree.IElementType;\n");
         sb.append("import iggy.ide.lang.IggyLang;\n");
         sb.append("\n");
         sb.append("import java.util.HashMap;\n");
-        sb.append("import java.util.IdentityHashMap;\n");
         sb.append("import java.util.Map;\n");
-        sb.append("import java.util.function.Function;\n");
         sb.append("\n");
         sb.append("public class " + className + "ElementTypes {\n");
         metaSymbolNodes.forEach(s -> sb.append(generateElementType(s)));
@@ -51,20 +43,11 @@ public class GenerateParserFiles extends Generator {
         generateTypes(sb, this::generateElementType);
         sb.append("\n");
         sb.append("    private static final Map<String, IElementType> stringToElementTypeMap = new HashMap<>();\n");
-        sb.append("    private static final Map<IElementType, Function<ASTNode, PsiElement>> elementTypeToPsiElementMap = new IdentityHashMap<>();\n");
         sb.append("\n");
         sb.append("    static {\n");
         metaSymbolNodes.forEach(s -> sb.append(generateStringToElementTypetMapEntry(s)));
         sb.append("\n");
-        metaSymbolNodes.forEach(s -> sb.append(generateTypeToPsiElementMapEntry(s)));
-        sb.append("\n");
         generateTypes(sb, this::generateStringToElementTypetMapEntry);
-        sb.append("\n");
-        generateTypes(sb, this::generateTypeToPsiElementMapEntry);
-        sb.append("    }\n");
-        sb.append("\n");
-        sb.append("    public static PsiElement createPsiElement(ASTNode node) {\n");
-        sb.append("        return elementTypeToPsiElementMap.get(node.getElementType()).apply(node);\n");
         sb.append("    }\n");
         sb.append("\n");
         sb.append("    public static IElementType getElementType(String name) {\n");
@@ -75,32 +58,8 @@ public class GenerateParserFiles extends Generator {
         writeToJavaFile(sb.toString(), parserGenDirectory, className + "ElementTypes");
     }
 
-    private void generateTypes(StringBuilder sb, Function<String, String> f) {
-        for (Map.Entry<Nonterminal, List<RuntimeRule>> entry : grammar.getDefinitions().entrySet()) {
-            String nonterminalName = entry.getKey().getName();
-            List<RuntimeRule> alternatives = entry.getValue();
-
-            if (alternatives.size() == 0) {
-                sb.append(f.apply(nonterminalName));
-            } else if (alternatives.size() == 1) {
-                sb.append(f.apply(nonterminalName));
-            } else {
-                for (RuntimeRule alternative : alternatives) {
-                    if (alternative.getLabel() == null)
-                        throw new RuntimeException("All alternatives must have a label: " + alternative);
-                    String nodeName = alternative.getLabel() + nonterminalName.substring(0, 1).toUpperCase() + nonterminalName.substring(1);
-                    sb.append(f.apply(nodeName));
-                }
-            }
-        }
-    }
-
     private String generateElementType(String type) {
         return "    public static IElementType " + type + " = new IggyElementType(\"" + type + "\", IggyLang.instance);\n";
-    }
-
-    private String generateTypeToPsiElementMapEntry(String type) {
-        return "        elementTypeToPsiElementMap.put(" + type + ", node -> new iggy.ide.psi.IggyPsiElement." + type + "(node));\n";
     }
 
     private String generateStringToElementTypetMapEntry(String type) {
